glm: Add version 0.9.7.1
A single unit test is patched to remove the strings "Error:" and
"error:" from its output.  This is because msbuild will automatically
fail any build containing these strings in its output, irrespective of
its exit status.  This is of course hideously broken, but we need to
work around it to build successfully.