Refactor to use absolute install paths
This unifies the way in which packages are build with make on
Unix and everything else (basically Windows).  We no longer use
a DESTDIR install when using make.  This makes the Unix build a
bit less flexible, but it does make this package simpler to
maintain and test.  It was also a requirement for adding glm,
and potentially other packages in the future.