jenkins-build: Add 'platform' or 'standalone' to bundle name
platform: No third-party prerequisites, built against platform
          libraries
standalone: Contains and is built against all needed third-
            party libraries

Previously, we had to store these in separate directories and the
fact that we had the different artefacts with the same name was
ambigious.  This makes it explicit what is contained, and also
simplifies the release jobs and download page generation.