Add tool-cache option
Allow packages to be classified as "source" or "tool".  This
permits the downloads to be put in separate locations since
the build tools aren't needed as sources for e.g. GPL
compliance and rebuilding.  This is essentially renaming the
python cache to tool cache to allow for more general use with
non-python tools such as doxygen and patch in addition to
sphinx and genshi.