docs: Add super-build documentation
Moved from the ome-files docs.  This adds the documentation as a
super-build package, so that it can depend upon the documentation
installed by other packages e.g. for relative doxygen and sphinx
HTML links with intersphinx in the future.

It can be built as part of the regular superbuild, or with make
for convenience.